Resolve autoload include issue with ppm file name
When ppm creates children it generates a stub php file that includes the autoload file; this references the phar archive via the phars stub mapping (Phar::mapPhar()). There can only be a single map name (usually <filename>.phar). That means all links into the phar have to be to the <filename>.phar otherwise they will not load. For consistency, the ppm file is now mv’d as ppm.phar and symlinked to ppm.